update styling of list view and refactor list view controllers
