+++ app/assets/stylesheets/comfy/admin/cms/custom.sass
@@ -18,7 +18,7 @@ select[id=layout_app_layout]
 // customizations for violet branding
 #comfy
 	.badge-primary, .btn-primary, .btn-success
-		background-color: #6F5AFE
+		background-color: var(--color-primary)
 		border: none
 	.badge-secondary, .btn-secondary
 		background-color: #A75AFE
